Split the test module in two.
